Connection.pm: &_do_safely: avoid a race condition by delaying the after callback.
There is a race condition in RemoteControl - if a client issues a
request of its own immediately after serving a request, then the other
side may not yet have prepared itself to be a server by reclaiming
it's own connection.  This results in the following:

    * client claims the selection, unaware that the server does not
      yet own it

    * server claims connection and waits, unaware that it took the
      selection away from the client

    * the client learns that it lost the selection and treats that as
      an acknowledgement, unaware that the server is inactive

Further behaviour is undefined.  If the client is Zircon it promptly
fails because it detects that the server did not read the request
before taking the selection.  If both sides continue then they remain
out of sync with the client being one step ahead of the server.  There
may be failures if one side tries to read the selection when the other
side is not ready to handle the request.  If neither side fails then
the client finishes before the server and the server then times out
waiting for the client to read its selection and reclaim it.