Z:T:Selection - remove waitVariable from ->own
The own call now returns after queuing the event which causes SelectionOwn,
so we cannot get stuck in a tangled waitVariable.

It should not affect semantics of code between call here & return to
event processing, because that code must already assume the message
takes time to arrive and have effect.