Add the copyright-notice file
As per our internal agreement regarding all HAVANA repositories, EBI
copyright only starts in the year we first made changes to the code
(i.e. this one).

Regarding the starting year of the GRL copyright, it is based on the
date of the first commit in Git history so it might be necessary to
extend this further into the past if said history is incomplete.